refactor: Mudançao do tipo em author service, buscar por id

+++ src/main/java/api/library/study_library/model/service/AuthorService.java
@@ -9,6 +9,6 @@ public interface AuthorService {
     Author salvarAuthor(Author author);
     Author atualizarAuthor(Author author);
     void deletarAuthor(Author author);
-    Optional<Author> buscarAuthor(Long id);
+    Optional<Author> buscarAuthor(Integer id);
     List<Author> listarAutores();
 }
